fix(field-accel-b2): fifth-review triage — zk lanes restored workspace-wide, fail-closed policies, one commit recipe
- bound module compiles in ALL feature lanes again (round 4's zk cfg-gate
  broke the mandated cargo clippy --all --features host,zk: examples
  import the module unconditionally and feature unification enables
  field-accel-prototype+zk together); the sidecar stays non-ZK by
  failing CLOSED at the bridge under zk-blinded commitments, and only
  the roundtrip tests are zk-gated (with rationale)
- inverse() policy settled fail-closed: None answers invertibility for
  canonical values; a non-canonical input spoils (a violated type
  invariant must not yield a provable None/zero branch)
- one commit_region_poly recipe shared by prove and commit_advice_region
  (kills the last duplicated commit spelling); region_word_count doc
  scoped to the sidecar (the main pipeline pads, we reject) and the
  count actually threaded through allocation and Dory context
- error precedence: blob-size before geometry; DIVQ host expect message
  scoped to raw-insn guests; PROTOTYPE.md canonicity family now names
  the silent-corruption add() member
- workspace clippy lanes (--all, host and host,zk) join the permanent
  gate list; that surfaced and fixed pre-existing debt: fqmul guest
  missing the field-inline feature declaration, needless u64 casts

All gates green by exit code: fmt, clippy --all both modes, field_accel
both modes, inlines, muldiv both modes, both e2e examples (6+1 gates).

+++ examples/aleo-transfer/src/main.rs
@@ -5,10 +5,10 @@ pub fn main() {
 
     // drain any log noise from prior compiles, then trace the transfer
     let _ = jolt_inlines_edwards_bls12::sequence_builder::take_field_op_log();
-    let summary = guest::analyze_transfer_private(0xA1E0_0001_u64 as u64);
+    let summary = guest::analyze_transfer_private(0xA1E0_0001_u64);
 
     let transfer_log = jolt_inlines_edwards_bls12::sequence_builder::take_field_op_log();
-    let usdcx_summary = guest::analyze_usdcx_transfer_private(0xA1E0_0003_u64 as u64);
+    let usdcx_summary = guest::analyze_usdcx_transfer_private(0xA1E0_0003_u64);
     println!(
         "usdcx_transfer_private trace: {} cycles",
         usdcx_summary.trace_len()
@@ -150,7 +150,7 @@ pub fn main() {
         );
     }
 
-    let soft = guest::analyze_usdcx_transfer_private_soft(0xA1E0_0003_u64 as u64);
+    let soft = guest::analyze_usdcx_transfer_private_soft(0xA1E0_0003_u64);
     println!(
         "usdcx_transfer_private_soft trace: {} cycles",
         soft.trace_len()
@@ -161,7 +161,7 @@ pub fn main() {
         let blob =
             jolt_inlines_edwards_bls12::sequence_builder::build_record_blob_padded(&usdcx_log);
         let bound_summary = guest::analyze_usdcx_transfer_private_bound(
-            0xA1E0_0003_u64 as u64,
+            0xA1E0_0003_u64,
             jolt_sdk::UntrustedAdvice::new(blob.as_slice()),
         );
         let _ = jolt_inlines_edwards_bls12::sequence_builder::take_field_op_log();
